Make this a general purpose tool and move the importer into a subcommand of its own.
